requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
PyGithub>=2.1.0
anthropic>=0.40.0
rich>=13.7.0
